8901: watch out for non-standard output from solve_lin_sys
solve_lin_sys will return keys like 1.0*A1 instead of A1 when floats are
involved. Not being sure whether these will always just be 1.0 I have
taken the safe route and just divided the corresponding value by the
coefficient and have affirmed that all that was left was a symbol.
Modification will be needed if this affirmation is not justified. I did
not modify solve_lin_sys to modify the generators since I am not familiar
enough with how DMP is supposed to work. e.g. One can create Poly(x+2,x+1)
and the generator is x + 1; or Poly(2*x + 1, 2*x) with generator a Mul...but
neither of these gives roots for the generator and neither gives the
coefficient of that generators

>>> Poly(x+1 + 1, x+1).nth(1)
0
>>> Poly(2*x + 1, 2*x).nth(1)
0
>>> roots(Poly(2*x + 1, 2*x))
{}
>>> roots(Poly(x + 1 + 1, x+1))
{}